Add Van der Waals EoS
Add Van der Waals EoS options for H₂, He, H₂O, CH₄, NH₃, and CO₂ gases. The results are reliable up to 0.1 GPa, which is comparable to the ideal gas model. The Van der Waals EoS underestimates the density, which helps avoid the unrealistic scenario of heavier materials lying on top of lighter materials, a potential issue with the ideal gas model. Additionally, a typo in the equation for dPdT_​S​ for the ideal gas has been corrected. Fortunately, this typo had not been used previously.