Core temperature-gradient algorithm modernized
For historical reasons, the  routine that sets the core temperature gradient had no access to  and had fallen far out of date. The routine has been rewritten: it now calls , runs faster, and is much cleaner. This update also fixes the error that occurred when calculating a pure gas planet. In addition, the initial density guess in  and in  has been tweaked for extra robustness.